Extract ObservationUnitPosition.PhenotypeSearchDTO
Reuse in Middleware ObservationUnitPatchRequestDTO

Issue: IBP-3213
Reviewer: None
